+++ app.js
@@ -77,6 +77,7 @@ import __fileRoutes from './mongoose/routes/fileRoutes.js';
 import __paperlessTags from './mongoose/config/paperlessTagsConfig.js';
 import __paperlessRoutes from './mongoose/routes/paperlessRoutes.js';
 import __bankRoutes from './mongoose/routes/bankRoutes.js';
+import __accountantRoutes from './mongoose/routes/accountantRoutes.js';
 import __overviewRoutes from './mongoose/routes/overviewRoutes.js';
 import __ssoRoutes from './mongoose/routes/ssoRoutes.js';
 import __helpRoutes from './mongoose/routes/helpRoutes.js';
@@ -538,6 +539,7 @@ const main = async () => {
     appRouter.use('/', __fileRoutes);
     appRouter.use('/', __paperlessRoutes);
     appRouter.use('/', __bankRoutes);
+    appRouter.use('/', __accountantRoutes);
     appRouter.use('/', __overviewRoutes);
     appRouter.use('/', __ssoRoutes);
     appRouter.use('/', __helpRoutes);